requirements.txt
pinecone-client==2.2.4
openai==1.3.7
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
numpy==1.26.2
pydantic==2.5.2
python-multipart==0.0.6
tiktoken==0.5.1
langchain==0.0.350
